refactor(config): split load versus keys
